Add Authentication to Manager pages
Access to Manager pages now requires authentication. This is a very
crude authentication implementation. The database stores a username,
salt, and hashed password/salt (SHA-256 on the password+salt is stored
as the password in the database). When the user logs in, it compares the
salt+provided password hashed to the hashed result in the database, and
if it matches, the Session collection is updated with variables for
permission level and UserName.